Implement queue system
This is a first attempt to implement a queue system.  The following
commands are implemented:

- hubot queue sandbox - adds /me to the queue
- hubot (de|un)queue sandbox - removes /me from the queue
- hubot remove <user> from queue - forcibly removes <user> from the
queue
- hubot clear queue - clears the queue
- hubot show queue - shows a list (in priority order) of who's in the
queue

When a user releases a sandbox, then the next person in the queue will
be auto-assigned that sandbox and the person will be removed from the
queue.

I am new to coffeescript, so it is likely I have done something dumb.  I
also can't get the queue to initialize/save properly , at least in my
test environment.  I always have to do "hubot clear queue" first to be
able to do anything with it.  I don't know if it is just the environment
or if it is a problem with the script.